refactor(api): migrate repository for /api/admin/users/{id}/organizations
